Widen tally filter-bin indexing to int64
n_filter_bins_ and strides_ were int32, so the product of per-filter
bin counts overflowed 2^31 for large tallies (a fine mesh crossed with
a fine energy filter can reach ~4e9 bins). Widen the filter-bin index
path to int64; per-filter n_bins stays int32 since only the products
overflow. No behavior change for tallies that fit int32 today. Adds a
Catch2 test for >2^31 stride/index math.

First step toward distributed tally storage (shared/rma modes) for
memory-bound mesh-energy tallies: nothing else in that effort can
allocate or address these tallies until the indexing is 64-bit.

+++ include/openmc/random_ray/source_region.h
@@ -51,10 +51,12 @@ inline void hash_combine(size_t& seed, const size_t v)
 // every iteration.
 struct TallyTask {
   int tally_idx;
-  int filter_idx;
+  // int64 to match the tally's int64 flat filter-bin index; a fine mesh
+  // crossed with a fine energy filter can exceed 2^31 combinations.
+  int64_t filter_idx;
   int score_idx;
   int score_type;
-  TallyTask(int tally_idx, int filter_idx, int score_idx, int score_type)
+  TallyTask(int tally_idx, int64_t filter_idx, int score_idx, int score_type)
     : tally_idx(tally_idx), filter_idx(filter_idx), score_idx(score_idx),
       score_type(score_type)
   {}
